no-api-news-scraper/crawler/requirements.txt
httpx==0.27.*
feedparser==6.0.*
pyyaml==6.0.*
playwright==1.48.*
pydantic==2.9.*
SQLAlchemy==2.0.*
click==8.1.*
pytest==7.4.*
